# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AuditEventRead": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "new_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "New Value"
          },
          "old_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Old Value"
          },
          "source": {
            "$ref": "#/components/schemas/AuditSource"
          }
        },
        "required": [
          "id",
          "action",
          "field",
          "old_value",
          "new_value",
          "source",
          "actor_user_id",
          "correlation_id",
          "created_at"
        ],
        "title": "AuditEventRead",
        "type": "object"
      },
      "AuditSource": {
        "enum": [
          "system",
          "accepted_suggestion",
          "override",
          "handler"
        ],
        "title": "AuditSource",
        "type": "string"
      },
      "CaseAssign": {
        "additionalProperties": false,
        "properties": {
          "assigned_to_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          }
        },
        "required": [
          "assigned_to_user_id"
        ],
        "title": "CaseAssign",
        "type": "object"
      },
      "CaseCreate": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "external_message_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Email Message-ID. Submitting the same id twice returns the existing case.",
            "title": "External Message Id"
          },
          "received_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the complaint was received. Defaults to now.",
            "title": "Received At"
          },
          "sender_email": {
            "format": "email",
            "title": "Sender Email",
            "type": "string"
          },
          "sender_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "subject": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "body",
          "sender_email"
        ],
        "title": "CaseCreate",
        "type": "object"
      },
      "CaseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CaseList",
        "type": "object"
      },
      "CaseRead": {
        "properties": {
          "assigned_to_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deadline_tracking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeadlineTrackingRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "external_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Message Id"
          },
          "final_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_response_deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Response Deadline At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "resolution_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the case needs a person, when its status says so",
            "title": "Review Reason"
          },
          "sender_email": {
            "title": "Sender Email",
            "type": "string"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "src_deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Src Deadline At"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "suggestion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SuggestionRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null until the analysis completes (US-2.4)"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vulnerability_indicators": {
            "items": {
              "$ref": "#/components/schemas/IndicatorRead"
            },
            "title": "Vulnerability Indicators",
            "type": "array"
          }
        },
        "required": [
          "id",
          "reference",
          "external_message_id",
          "subject",
          "body",
          "sender_email",
          "sender_name",
          "received_at",
          "status",
          "assigned_to_user_id",
          "src_deadline_at",
          "final_response_deadline_at",
          "final_category",
          "final_priority",
          "resolved_at",
          "resolution_type",
          "created_at",
          "updated_at"
        ],
        "title": "CaseRead",
        "type": "object"
      },
      "CaseStatus": {
        "enum": [
          "new",
          "analysing",
          "awaiting_review",
          "needs_human_review",
          "in_progress",
          "resolved"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "Category": {
        "enum": [
          "fees_and_charges",
          "service_quality",
          "account_administration",
          "payments_and_transfers",
          "lending_and_affordability",
          "fraud_and_scams",
          "advice_and_mis_selling",
          "collections_and_arrears",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "CategoryDecision": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "required": [
          "category"
        ],
        "title": "CategoryDecision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndicatorCreate": {
        "additionalProperties": false,
        "description": "A vulnerability indicator the AI missed. The quote must be in the complaint.",
        "properties": {
          "evidence_quote": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Evidence Quote",
            "type": "string"
          },
          "indicator_type": {
            "$ref": "#/components/schemas/IndicatorType"
          }
        },
        "required": [
          "indicator_type",
          "evidence_quote"
        ],
        "title": "IndicatorCreate",
        "type": "object"
      },
...
      "IndicatorDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "confirmed",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "IndicatorDecisionRequest",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "Priority",
        "type": "string"
      },
      "PriorityDecision": {
        "additionalProperties": false,
        "properties": {
          "priority": {
            "$ref": "#/components/schemas/Priority"
          }
        },
        "required": [
          "priority"
        ],
        "title": "PriorityDecision",
        "type": "object"
      },
      "ResolutionType": {
        "enum": [
          "src",
          "final_response"
        ],
        "title": "ResolutionType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-assisted intake for regulated complaints. All data is fictional.",
    "title": "Complaint Intake Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cases": {
      "get": {
        "operationId": "list_cases_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Cases",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Duplicate: existing case returned"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}": {
      "get": {
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/assignment": {
      "patch": {
        "operationId": "assign_case_cases__case_id__assignment_patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/audit": {
      "get": {
        "operationId": "get_case_audit_cases__case_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventRead"
                  },
                  "title": "Response Get Case Audit Cases  Case Id  Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Case Audit",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/category": {
      "put": {
        "description": "Accept the suggested category (send the same value) or override it (send another).",
        "operationId": "decide_category_cases__case_id__category_put",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Category",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/priority": {
      "put": {
        "description": "Accept the suggested priority (send the same value) or override it (send another).",
        "operationId": "decide_priority_cases__case_id__priority_put",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Priority",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/vulnerability-indicators": {
      "post": {
        "description": "Add an indicator the AI missed, with a quote from the complaint as evidence (US-3.5).",
        "operationId": "add_indicator_cases__case_id__vulnerability_indicators_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndicatorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Indicator",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/vulnerability-indicators/{indicator_id}/decision": {
      "put": {
        "description": "Confirm or reject an indicator. A rejected one stays visible as rejected (US-3.4).",
        "operationId": "decide_indicator_cases__case_id__vulnerability_indicators__indicator_id__decision_put",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "indicator_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Indicator Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndicatorDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Indicator",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}